app/models/inference.py
"""
ML Inference module for underwater cable analysis
Integrates with Ollama and other ML frameworks
"""

import ollama
from PIL import Image
import io
import base64
from typing import Dict, List, Optional
import numpy as np


class CableAnalysisModel:
    """
    Main model class for underwater cable image analysis
    Supports Ollama vision models and custom ML models
    """

    def __init__(self, model_name: str = "llama3.2-vision:11b"):
        """
        Initialize the analysis model

        Args:
            model_name: Name of the Ollama model to use for vision analysis
        """
        self.model_name = model_name
        self.analysis_prompt = """
        You are an expert in underwater cable inspection and maintenance.
        Analyze this underwater cable image and provide:

        1. Overall cable condition (excellent, good, fair, poor, critical)
        2. Specific issues detected (corrosion, physical damage, wear, biological growth, etc.)
        3. Severity level for each issue (low, medium, high, critical)
        4. Maintenance recommendations

        Be specific and concise in your analysis.
        """

    async def analyze_image(self, image_path: str) -> Dict:
        """
        Analyze an underwater cable image using Ollama vision model

        Args:
            image_path: Path to the image file

        Returns:
            Dictionary containing analysis results
        """
        try:
            # Use Ollama for vision-based analysis
            response = ollama.chat(
                model=self.model_name,
                messages=[
                    {
                        'role': 'user',
                        'content': self.analysis_prompt,
                        'images': [image_path]
                    }
                ]
            )

            # Parse the response
            analysis_text = response['message']['content']

            # Extract structured information from the response
            result = self._parse_analysis_response(analysis_text)

            return result

        except Exception as e:
            return {
                "error": str(e),
                "status": "failed",
                "message": "Failed to analyze image with ML model"
            }

    def _parse_analysis_response(self, analysis_text: str) -> Dict:
        """
        Parse the LLM response into structured data

        Args:
            analysis_text: Raw text response from the model

        Returns:
            Structured analysis results
        """
        # Basic parsing - can be enhanced with more sophisticated extraction
        result = {
            "raw_analysis": analysis_text,
            "detected_issues": [],
            "cable_condition": "unknown",
            "confidence_score": 0.75,  # Placeholder
            "recommendations": []
        }

        # Extract condition
        text_lower = analysis_text.lower()
        if "excellent" in text_lower:
            result["cable_condition"] = "excellent"
        elif "good" in text_lower:
            result["cable_condition"] = "good"
        elif "fair" in text_lower:
            result["cable_condition"] = "fair"
        elif "poor" in text_lower:
            result["cable_condition"] = "poor"
        elif "critical" in text_lower:
            result["cable_condition"] = "critical"

        # Extract common issues
        issues = [
            "corrosion", "damage", "wear", "biological growth",
            "crack", "tear", "degradation", "fouling"
        ]

        for issue in issues:
            if issue in text_lower:
                result["detected_issues"].append(issue)

        # Extract recommendations (lines starting with common indicators)
        lines = analysis_text.split('\n')
        for line in lines:
            line_lower = line.lower().strip()
            if any(starter in line_lower for starter in ["recommend", "should", "suggest", "inspect", "replace", "repair"]):
                if line.strip():
                    result["recommendations"].append(line.strip())

        return result

    async def analyze_with_custom_model(self, image_path: str) -> Dict:
        """
        Placeholder for custom ML model integration
        (e.g., PyTorch, TensorFlow, ONNX models)

        Args:
            image_path: Path to the image file

        Returns:
            Dictionary containing analysis results
        """
        # TODO: Implement custom model inference
        # This could integrate models trained specifically for cable defect detection

        return {
            "status": "not_implemented",
            "message": "Custom model integration pending"
        }

    def preprocess_image(self, image_path: str, target_size: tuple = (224, 224)) -> np.ndarray:
        """
        Preprocess image for ML model input

        Args:
            image_path: Path to the image file
            target_size: Target size for the image

        Returns:
            Preprocessed image as numpy array
        """
        img = Image.open(image_path)
        img = img.resize(target_size)
        img_array = np.array(img) / 255.0  # Normalize to [0, 1]
        return img_array
